backend/requirements.txt
Flask
Flask-Cors
google-generativeai
Pillow
python-dotenv
opencv-python
mediapipe
